chore: introduce deferred terminology
