+++ src/sdk/clients/createNexusSessionClient.test.ts
@@ -129,10 +129,9 @@ describe("nexus.session.client", async () => {
 
     nexusClient.account.getCounterFactualAddress()
 
-    const createSessionsResponse =
-      await nexusSessionClient.grantPermissionInAdvance({
-        sessionRequestedInfo
-      })
+    const createSessionsResponse = await nexusSessionClient.grantPermission({
+      sessionRequestedInfo
+    })
 
     expect(createSessionsResponse.userOpHash).toBeDefined()
     expect(createSessionsResponse.permissionIds).toBeDefined()
