+++ src/sdk/modules/smartSessionsValidator/decorators/grantPermission.ts
@@ -1,14 +1,10 @@
-import type { Chain, Client, PublicClient, Transport } from "viem"
+import type { Chain, Client, Hex, PublicClient, Transport } from "viem"
+import { sendUserOperation } from "viem/account-abstraction"
 import { getAction, parseAccount } from "viem/utils"
 import { AccountNotFoundError } from "../../../account/utils/AccountNotFound"
-import {
-  MAINNET_ADDRESS_K1_VALIDATOR_ADDRESS,
-  SmartSessionMode,
-  getAccount,
-  getEnableSessionDetails
-} from "../../../constants"
+import type { Call } from "../../../account/utils/Types"
 import type { ModularSmartAccount } from "../../utils/Types"
-import type { CreateSessionDataParams, SessionData } from "../Types"
+import type { CreateSessionDataParams, GrantPermissionResponse } from "../Types"
 import { preparePermission } from "./preparePermission"
 
 /**
@@ -21,13 +17,22 @@ export type GrantPermissionParameters<
 > = {
   /** Array of session data parameters for creating multiple sessions. */
   sessionRequestedInfo: CreateSessionDataParams[]
+  /** The maximum fee per gas unit the transaction is willing to pay. */
+  maxFeePerGas?: bigint
+  /** The maximum priority fee per gas unit the transaction is willing to pay. */
+  maxPriorityFeePerGas?: bigint
+  /** The nonce of the transaction. If not provided, it will be determined automatically. */
+  nonce?: bigint
   /** Optional public client for blockchain interactions. */
   publicClient?: PublicClient
   /** The modular smart account to create sessions for. If not provided, the client's account will be used. */
   account?: TModularSmartAccount
+  /** Optional attesters to trust. */
+  attesters?: Hex[]
+  /** Additional calls to be included in the user operation. */
+  calls?: Call[]
 }
 
-export type GrantPermissionResponse = SessionData["moduleData"]
 /**
  * Adds multiple sessions to the SmartSessionValidator module of a given smart account.
  *
@@ -78,7 +83,13 @@ export async function grantPermission<
   client: Client<Transport, Chain | undefined, TModularSmartAccount>,
   parameters: GrantPermissionParameters<TModularSmartAccount>
 ): Promise<GrantPermissionResponse> {
-  const { account: account_ = client.account } = parameters
+  const {
+    account: account_ = client.account,
+    maxFeePerGas,
+    maxPriorityFeePerGas,
+    nonce,
+    calls: calls_
+  } = parameters
 
   if (!account_) {
     throw new AccountNotFoundError({
@@ -87,8 +98,6 @@ export async function grantPermission<
   }
 
   const account = parseAccount(account_) as ModularSmartAccount
-  const publicClient = account?.client as PublicClient
-
   if (!account || !account.address) {
     throw new Error("Account not found")
   }
@@ -99,30 +108,26 @@ export async function grantPermission<
     "preparePermission"
   )(parameters)
 
-  const nexusAccount = getAccount({
-    address: account.address,
-    type: "nexus"
+  const userOpHash = await getAction(
+    client,
+    sendUserOperation,
+    "sendUserOperation"
+  )({
+    calls: [
+      {
+        to: preparedPermission.action.target,
+        data: preparedPermission.action.callData
+      },
+      ...(calls_ || [])
+    ],
+    maxFeePerGas,
+    maxPriorityFeePerGas,
+    nonce,
+    account
   })
-
-  const sessionDetailsWithPermissionEnableHash = await getEnableSessionDetails({
-    enableMode: SmartSessionMode.UNSAFE_ENABLE,
-    sessions: preparedPermission.sessions,
-    account: nexusAccount,
-    clients: [publicClient],
-    enableValidatorAddress: MAINNET_ADDRESS_K1_VALIDATOR_ADDRESS
-  })
-
-  const { permissionEnableHash, ...sessionDetails } =
-    sessionDetailsWithPermissionEnableHash
-
-  sessionDetails.enableSessionData.enableSession.permissionEnableSig =
-    await account.signer.signMessage({ message: { raw: permissionEnableHash } })
 
   return {
-    permissionIds: preparedPermission.permissionIds,
-    action: preparedPermission.action,
-    mode: SmartSessionMode.UNSAFE_ENABLE,
-    sessions: preparedPermission.sessions,
-    enableSessionData: sessionDetails.enableSessionData
+    userOpHash,
+    ...preparedPermission
   }
 }
