+++ src/sdk/modules/smartSessionsValidator/toSmartSessionValidator.enable.mode.test.ts
@@ -248,7 +248,7 @@ describe("modules.smartSessions.enable.mode.dx", async () => {
       smartSessionCreateActions(sessionsModule)
     )
 
-    const moduleData = await nexusSessionClient.grantPermission({
+    const moduleData = await nexusSessionClient.grantDeferredPermission({
       sessionRequestedInfo: [
         {
           sessionPublicKey, // Public key of the session
