+++ src/sdk/modules/smartSessionsValidator/toSmartSessionsValidator.policies.test.ts
@@ -113,32 +113,31 @@ describe("modules.smartSessions.policies", async () => {
 
     // Define the session parameters
     // This includes the session key, validator, and action policies
-    const createSessionsResponse =
-      await nexusSessionClient.grantPermissionInAdvance({
-        sessionRequestedInfo: [
-          {
-            sessionPublicKey, // Public key of the session
-            sessionValidUntil: Date.now() + 1000 * 60 * 60 * 24, // 1 day from now
-            chainIds: [BigInt(chain.id)],
-            actionPoliciesInfo: [
-              {
-                contractAddress: testAddresses.Counter,
-                sudo: false, // covered in another test
-                tokenLimits: [
-                  {
-                    token: "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2", // WETH
-                    limit: BigInt(1000)
-                  }
-                ], // covered in another test
-                // usageLimit: 1000n, // TODO: failing because of attestations
-                // valueLimit: 1000n, // TODO: failing because of attestations
-                validUntil: Date.now() + 1000 * 60 * 60 * 24, // 1 day from now
-                functionSelector: "0x871cc9d4" // decrementNumber
-              }
-            ]
-          }
-        ]
-      })
+    const createSessionsResponse = await nexusSessionClient.grantPermission({
+      sessionRequestedInfo: [
+        {
+          sessionPublicKey, // Public key of the session
+          sessionValidUntil: Date.now() + 1000 * 60 * 60 * 24, // 1 day from now
+          chainIds: [BigInt(chain.id)],
+          actionPoliciesInfo: [
+            {
+              contractAddress: testAddresses.Counter,
+              sudo: false, // covered in another test
+              tokenLimits: [
+                {
+                  token: "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2", // WETH
+                  limit: BigInt(1000)
+                }
+              ], // covered in another test
+              // usageLimit: 1000n, // TODO: failing because of attestations
+              // valueLimit: 1000n, // TODO: failing because of attestations
+              validUntil: Date.now() + 1000 * 60 * 60 * 24, // 1 day from now
+              functionSelector: "0x871cc9d4" // decrementNumber
+            }
+          ]
+        }
+      ]
+    })
 
     // Wait for the session creation transaction to be mined and check its success
     const { success: sessionCreateSuccess } =
