+++ src/test/playground.test.ts
@@ -214,10 +214,9 @@ describe.skipIf(!playgroundTrue())("playground", () => {
       smartSessionCreateActions(sessionsModule)
     )
 
-    const createSessionsResponse =
-      await nexusSessionClient.grantPermissionInAdvance({
-        sessionRequestedInfo
-      })
+    const createSessionsResponse = await nexusSessionClient.grantPermission({
+      sessionRequestedInfo
+    })
 
     expect(createSessionsResponse.userOpHash).toBeDefined()
     expect(createSessionsResponse.permissionIds).toBeDefined()
